+++ SharePointFramework/ProjectWebParts/src/components/ProjectStatus/Sections/BaseSection/BaseSection.module.scss
@@ -1,12 +1,13 @@
 @import "~@microsoft/sp-office-ui-fabric-core/dist/sass/SPFabricCore.scss";
 
 .root {
+  @include ms-Grid;
   margin: 30px 0 0 0;
   background-color: #eff6fa;
   border-top: 5px solid #035484;
-  @include ms-Grid;
+
   .container {
-    padding: 15px 30px 30px 30px;
     @include ms-Grid-row;
+    padding: 15px 30px 30px 30px;
   }
 }
