+++ SharePointFramework/ProjectWebParts/src/components/ProjectStatus/Sections/ProjectPropertiesSection/StatusSectionField/StatusSectionField.module.scss
@@ -1,5 +1,6 @@
 .root {
   padding: 0 18px 18px 0;
+  box-sizing: border-box;
   .container {
     .fieldLabel {
       font-size: 14px;
