+++ SharePointFramework/ProjectWebParts/src/components/ProjectStatus/Sections/StatusSection/index.tsx
@@ -1,8 +1,11 @@
-import React, { FC } from 'react'
+import React, { FC, useContext } from 'react'
 import { StatusElement } from '../../StatusElement'
 import { BaseSection } from '../BaseSection'
+import { SectionContext } from '../context'
 
 export const StatusSection: FC = () => {
+  const { headerProps } = useContext(SectionContext)
+  if (!headerProps.value) return null
   return (
     <BaseSection>
       <div className='ms-Grid-row'>
