+++ SharePointFramework/ProjectWebParts/src/components/ProjectStatus/StatusElement/index.tsx
@@ -9,16 +9,16 @@ export const StatusElement: FC = () => {
     <div className={styles.root}>
       <div className={styles.container}>
         <div
-          className={styles.statusIcon}
+          className={styles.icon}
           style={{ fontSize: headerProps.iconSize, color: headerProps.iconColor }}>
           <Icon iconName={headerProps.iconName} />
         </div>
-        <div className={styles.statusContent}>
-          <div className={styles.statusElementLabel}>{headerProps.label}</div>
-          <div className={styles.statusElementValue}>{headerProps.value}</div>
+        <div className={styles.content}>
+          <div className={styles.label}>{headerProps.label}</div>
+          <div className={styles.value}>{headerProps.value}</div>
           {headerProps.comment && (
             <div
-              className={styles.statusElementComment}
+              className={styles.comment}
               dangerouslySetInnerHTML={{
                 __html: headerProps.comment.replace(/\n/g, '<br />')
               }}></div>
