+++ SharePointFramework/ProjectWebParts/src/components/ProjectStatus/StatusElement/types.ts
@@ -1,14 +1,3 @@
-export type StatusElementColumnWidth =
-  | 'column1'
-  | 'column2'
-  | 'column3'
-  | 'column4'
-  | 'column6'
-  | 'column8'
-  | 'column10'
-  | 'column11'
-  | 'column12'
-
 export interface IStatusElement {
   label: string
   value: string
@@ -17,6 +6,4 @@ export interface IStatusElement {
   height?: string | number
   iconSize?: number
   iconColor?: string
-  iconColumnWidth?: StatusElementColumnWidth
-  bodyColumnWidth?: StatusElementColumnWidth
 }
